Make groups deal in fingerprints
Currently, alert groups retain a list of their alerts inside
themselves. This is superflous when all they really need is the
fingerprint and the state.

This change refactors things to make Groups be a `DehydratedAlert`
instead of a full alert. Dehydrated alerts have only a fingerprint
and a state, which makes groups a lot smaller.

Signed-off-by: sinkingpoint <colin@quirl.co.nz>