Support storing alert groups in the alert group controller
In order to support the GET /api/v2/groups endpoint, we need to store
alert groups somewhere. We _could_ do this by fanning out to each of the
alert-group-controller DOs, but in high throughput environments that
probably has the potential to be slow, so this adds in a centralised
cache in the account controller for now, and we can investigate
removing it later.

Signed-off-by: sinkingpoint <colin@quirl.co.nz>